requirements.txt
jax~=0.4.4
jax-spcs-kinematics
matplotlib~=3.7.0
numpy~=1.21.6
PyYAML~=6.0
pyvista~=0.35.2
seaborn~=0.11.2
natsort~=8.2.0
scipy~=1.10.1
pandas~=1.4.3